ui/requirements.txt
streamlit==1.41.1
openai==1.60.2
snowflake-connector-python
snowflake-snowpark-python
snowflake-ml-python